Oceanic : revert to simpler byte-driven serial reader thread [After moving to a linux based OH server, suddenly the stream of characters received from the Oceanic device contains 0x00 (‘\0’) bytes. I am not sure whether this is due to the serial drivers on Linux, the JVM used (openJDK) or something else]
