+++ addons/binding/org.openhab.binding.oceanic/src/main/java/org/openhab/binding/oceanic/OceanicBindingConstants.java
@@ -215,15 +215,15 @@ public class OceanicBindingConstants {
             @Override
             public String convertValue(String value) {
 
-                final SimpleDateFormat IN_DATE_FORMATTER = new SimpleDateFormat("dd.MM.yy HH:mm:ss");
-                final SimpleDateFormat OUT_DATE_FORMATTER = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
+                final SimpleDateFormat inDateFormatter = new SimpleDateFormat("dd.MM.yy HH:mm:ss");
+                final SimpleDateFormat outDateFormatter = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss");
 
                 Date date = null;
                 String convertedValue = null;
 
                 try {
-                    date = IN_DATE_FORMATTER.parse(value);
-                    convertedValue = OUT_DATE_FORMATTER.format(date);
+                    date = inDateFormatter.parse(value);
+                    convertedValue = outDateFormatter.format(date);
                 } catch (ParseException fpe) {
                     throw new IllegalArgumentException(value + " is not in a valid format.", fpe);
                 }
