+++ addons/binding/org.openhab.binding.oceanic/src/main/java/org/openhab/binding/oceanic/handler/OceanicThingHandler.java
@@ -8,18 +8,27 @@
  */
 package org.openhab.binding.oceanic.handler;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.math.BigDecimal;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Enumeration;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 
+import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.StringUtils;
 import org.eclipse.smarthome.core.thing.Channel;
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingStatus;
+import org.eclipse.smarthome.core.thing.ThingStatusDetail;
+import org.eclipse.smarthome.core.thing.binding.BaseThingHandler;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
 import org.eclipse.smarthome.core.types.State;
@@ -30,37 +39,42 @@ import org.openhab.binding.oceanic.internal.SerialPortThrottler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import gnu.io.CommPortIdentifier;
+import gnu.io.PortInUseException;
+import gnu.io.SerialPort;
+import gnu.io.UnsupportedCommOperationException;
+
 /**
  * The {@link OceanicHandler} is responsible for handling commands, which are
  * sent to one of the channels.
  *
  * @author Karel Goderis - Initial contribution
  */
-public class OceanicThingHandler extends SerialThingHandler {
+public class OceanicThingHandler extends BaseThingHandler {
 
     // List of Configuration constants
     public static final String INTERVAL = "interval";
+    public static final String PORT = "port";
+    public static final String BAUD_RATE = "baud";
+    public static final String BUFFER_SIZE = "buffer";
 
-    private Logger logger = LoggerFactory.getLogger(OceanicThingHandler.class);
+    private static Logger logger = LoggerFactory.getLogger(OceanicThingHandler.class);
+
+    private SerialPort serialPort;
+    private CommPortIdentifier portId;
+    private InputStream inputStream;
+    private OutputStream outputStream;
+    private int baud;
+    private String port;
+    private int bufferSize;
 
     private ScheduledFuture<?> pollingJob;
-    private String lastLineReceived = "";
-    private long GRACE_PERIOD = 1000;
-    private long REQUEST_TIMEOUT = 15000;
-    private long previousRequestTypeCommand = 0;
+    private SerialPortReader readerThread = null;
+    private static String lastLineReceived = "";
+    private long REQUEST_TIMEOUT = 10000;
 
     public OceanicThingHandler(Thing thing) {
         super(thing);
-    }
-
-    @Override
-    public void dispose() {
-        logger.debug("Disposing Oceanic handler.");
-        if (pollingJob != null && !pollingJob.isCancelled()) {
-            pollingJob.cancel(true);
-            pollingJob = null;
-        }
-        super.dispose();
     }
 
     @Override
@@ -81,53 +95,102 @@ public class OceanicThingHandler extends SerialThingHandler {
 
         port = (String) getConfig().get(PORT);
 
-        sleep = 250;
+        if (serialPort == null && port != null && baud != 0) {
 
-        super.initialize();
-
-        onUpdate();
-    }
-
-    private synchronized void onUpdate() {
-        if (pollingJob == null || pollingJob.isCancelled()) {
-            int polling_interval = ((BigDecimal) getConfig().get(INTERVAL)).intValue();
-            pollingJob = scheduler.scheduleWithFixedDelay(pollingRunnable, 1, polling_interval, TimeUnit.SECONDS);
-        }
-
-        // Read out the properties
-        try {
-            if (getThing().getStatus() == ThingStatus.ONLINE) {
-                for (Channel aChannel : getThing().getChannels()) {
-                    try {
-                        OceanicChannelSelector selector = OceanicChannelSelector.getValueSelector(
-                                aChannel.getUID().getId(), OceanicChannelSelector.ValueSelectorType.GET);
-
-                        if (selector != null && selector.isProperty()) {
-                            scheduler.schedule(new RequestRunnable(aChannel.getUID(), selector), 0,
-                                    TimeUnit.MILLISECONDS);
-                        }
-                    } catch (Exception e) {
+            @SuppressWarnings("rawtypes")
+            Enumeration portList = CommPortIdentifier.getPortIdentifiers();
+            while (portList.hasMoreElements()) {
+                CommPortIdentifier id = (CommPortIdentifier) portList.nextElement();
+                if (id.getPortType() == CommPortIdentifier.PORT_SERIAL) {
+                    if (id.getName().equals(port)) {
+                        logger.debug("Serial port '{}' has been found.", port);
+                        portId = id;
                     }
                 }
             }
-        } catch (Exception e) {
-            logger.error("An exception occurred while reading the properties the Oceanic Water Softener: '{}'",
-                    e.getMessage());
+
+            if (portId != null) {
+                try {
+                    serialPort = portId.open(this.getThing().getUID().getBindingId(), 2000);
+                } catch (PortInUseException e) {
+                    updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR,
+                            "Could not open serial port " + serialPort + ": " + e.getMessage());
+                    return;
+                }
+
+                try {
+                    inputStream = serialPort.getInputStream();
+                } catch (IOException e) {
+                    updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR,
+                            "Could not open serial port " + serialPort + ": " + e.getMessage());
+                    return;
+                }
+
+                serialPort.notifyOnDataAvailable(true);
+
+                try {
+                    serialPort.setSerialPortParams(baud, SerialPort.DATABITS_8, SerialPort.STOPBITS_1,
+                            SerialPort.PARITY_NONE);
+                } catch (UnsupportedCommOperationException e) {
+                    updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR,
+                            "Could not configure serial port " + serialPort + ": " + e.getMessage());
+                    return;
+                }
+
+                try {
+                    outputStream = serialPort.getOutputStream();
+                    updateStatus(ThingStatus.ONLINE);
+                } catch (IOException e) {
+                    updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR,
+                            "Could not communicate with the serial port " + serialPort + ": " + e.getMessage());
+                    return;
+                }
+
+                readerThread = new SerialPortReader(inputStream);
+                readerThread.start();
+            } else {
+                StringBuilder sb = new StringBuilder();
+                portList = CommPortIdentifier.getPortIdentifiers();
+                while (portList.hasMoreElements()) {
+                    CommPortIdentifier id = (CommPortIdentifier) portList.nextElement();
+                    if (id.getPortType() == CommPortIdentifier.PORT_SERIAL) {
+                        sb.append(id.getName() + "\n");
+                    }
+                }
+                logger.error("Serial port '" + port + "' could not be found. Available ports are:\n" + sb.toString());
+                updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR);
+            }
+        }
+
+        if (pollingJob == null || pollingJob.isCancelled()) {
+            pollingJob = scheduler.scheduleWithFixedDelay(pollingRunnable, 1,
+                    ((BigDecimal) getConfig().get(INTERVAL)).intValue(), TimeUnit.SECONDS);
         }
     }
 
     @Override
-    public void onDataReceived(String line) {
+    public void dispose() {
+        logger.debug("Disposing Oceanic handler.");
+        if (pollingJob != null && !pollingJob.isCancelled()) {
+            pollingJob.cancel(true);
+            pollingJob = null;
+        }
 
-        logger.trace("Received '{}'", line);
+        IOUtils.closeQuietly(inputStream);
+        IOUtils.closeQuietly(outputStream);
+        if (serialPort != null) {
+            serialPort.close();
+        }
 
-        line = StringUtils.chomp(line);
-
-        // little hack to overcome Locale limits of the Oceanic device
-        line = line.replace(",", ".");
-        line = line.trim();
-
-        lastLineReceived = line;
+        if (readerThread != null) {
+            try {
+                readerThread.interrupt();
+                readerThread.join();
+            } catch (InterruptedException e) {
+                logger.error("An exception occured while interrupting the serial port reader thread : '{}'",
+                        e.getMessage(), e);
+            }
+        }
     }
 
     private Runnable pollingRunnable = new Runnable() {
@@ -136,23 +199,27 @@ public class OceanicThingHandler extends SerialThingHandler {
         public void run() {
             try {
                 if (getThing().getStatus() == ThingStatus.ONLINE) {
-
-                    long interval = (((BigDecimal) getConfig().get(INTERVAL)).intValue() * 1000)
-                            / getThing().getChannels().size();
-                    logger.trace("Sending a request every {} milliseconds", interval);
-                    long delay = 0;
-
                     for (Channel aChannel : getThing().getChannels()) {
-                        try {
-                            OceanicChannelSelector selector = OceanicChannelSelector.getValueSelector(
-                                    aChannel.getUID().getId(), OceanicChannelSelector.ValueSelectorType.GET);
-
-                            if (selector != null) {
-                                scheduler.schedule(new RequestRunnable(aChannel.getUID(), selector), delay,
-                                        TimeUnit.MILLISECONDS);
-                                delay = delay + interval;
+                        for (OceanicChannelSelector selector : OceanicChannelSelector.values()) {
+                            ChannelUID theChannelUID = new ChannelUID(getThing().getUID(), selector.toString());
+                            if (aChannel.getUID().equals(theChannelUID)
+                                    && selector.getTypeValue() == OceanicChannelSelector.ValueSelectorType.GET) {
+                                String response = requestResponse(selector.name());
+                                if (response != "") {
+                                    if (selector.isProperty()) {
+                                        logger.debug("Updating the property '{}' with value '{}'", selector.toString(),
+                                                selector.convertValue(response));
+                                        Map<String, String> properties = editProperties();
+                                        properties.put(selector.toString(), selector.convertValue(response));
+                                        updateProperties(properties);
+                                    } else {
+                                        State value = createStateForType(selector, response);
+                                        updateState(theChannelUID, value);
+                                    }
+                                } else {
+                                    logger.warn("Received an empty answer for '{}'", selector.name());
+                                }
                             }
-                        } catch (Exception e) {
                         }
                     }
                 }
@@ -162,92 +229,40 @@ public class OceanicThingHandler extends SerialThingHandler {
         }
     };
 
-    private class RequestRunnable implements Runnable {
-
-        private OceanicChannelSelector selector;
-        private ChannelUID channelUID;
-
-        public RequestRunnable(ChannelUID theChannelUID, OceanicChannelSelector selector) {
-            this.selector = selector;
-            this.channelUID = theChannelUID;
-        }
-
-        @Override
-        public void run() {
-            try {
-                if (getThing().getStatus() == ThingStatus.ONLINE) {
-                    logger.trace("Requesting a response for {}", selector.name());
-                    String response = requestResponse(selector.name());
-                    if (response != "") {
-                        if (selector.isProperty()) {
-                            logger.debug("Updating the property '{}' with value '{}'", selector.toString(),
-                                    selector.convertValue(response));
-                            Map<String, String> properties = editProperties();
-                            properties.put(selector.toString(), selector.convertValue(response));
-                            updateProperties(properties);
-                        } else {
-                            State value = createStateForType(selector, response);
-                            updateState(channelUID, value);
-                        }
-                    } else {
-                        logger.warn("Received an empty answer for '{}'", selector.name());
-                    }
-                }
-            } catch (Exception e) {
-                logger.error("An exception occurred while requesting '{}' from the Oceanic Water Softener: '{}'",
-                        selector.toString(), e.getMessage());
-            }
-        }
-
-    }
-
     @Override
     public void handleCommand(ChannelUID channelUID, Command command) {
 
         if (getThing().getStatus() == ThingStatus.ONLINE) {
-
-            if (command instanceof RefreshType) {
-
+            if (!(command instanceof RefreshType)) {
+                String commandAsString = command.toString();
                 String channelID = channelUID.getId();
-                OceanicChannelSelector selector = OceanicChannelSelector.getValueSelector(channelID,
-                        OceanicChannelSelector.ValueSelectorType.GET);
 
-                if (selector != null) {
-                    scheduler.schedule(new RequestRunnable(channelUID, selector), 0, TimeUnit.MILLISECONDS);
-                }
+                for (Channel aChannel : getThing().getChannels()) {
+                    if (aChannel.getUID().equals(channelUID)) {
+                        try {
+                            OceanicChannelSelector selector = OceanicChannelSelector.getValueSelector(channelID,
+                                    OceanicChannelSelector.ValueSelectorType.SET);
 
-                return;
-            }
-
-            String commandAsString = command.toString();
-            String channelID = channelUID.getId();
-
-            for (Channel aChannel : getThing().getChannels()) {
-                if (aChannel.getUID().equals(channelUID)) {
-                    try {
-                        OceanicChannelSelector selector = OceanicChannelSelector.getValueSelector(channelID,
-                                OceanicChannelSelector.ValueSelectorType.SET);
-
-                        switch (selector) {
-                            case setSV1:
-                                commandAsString = selector.name() + commandAsString;
-                                break;
-                            default:
-                                commandAsString = selector.name();
-                                break;
+                            switch (selector) {
+                                case setSV1:
+                                    commandAsString = selector.name() + commandAsString;
+                                    break;
+                                default:
+                                    commandAsString = selector.name();
+                                    break;
+                            }
+                            String response = requestResponse(commandAsString);
+                            if (response.equals("ERR")) {
+                                logger.error("An error occurred while setting '{}' to {}", selector.toString(),
+                                        commandAsString);
+                            }
+                        } catch (IllegalArgumentException e) {
+                            logger.warn(
+                                    "An error occurred while trying to set the read-only variable associated with channel '{}' to '{}'",
+                                    channelID, command.toString());
                         }
-                        String response = requestResponse(commandAsString);
-                        if (response.equals("ERR")) {
-                            logger.error("An error occurred while setting '{}' to {}", selector.toString(),
-                                    commandAsString);
-                        }
-                    } catch (IllegalArgumentException e) {
-                        logger.warn(
-                                "An error occurred while trying to set the read-only variable associated with channel '{}' to '{}'",
-                                channelID, command.toString());
+                        break;
                     }
-
-                    break;
                 }
             }
         }
@@ -267,12 +282,11 @@ public class OceanicThingHandler extends SerialThingHandler {
     }
 
     private String requestResponse(String commandAsString) {
+        synchronized (this) {
+            SerialPortThrottler.lock(port);
 
-        SerialPortThrottler.lock(port);
-
-        String response = null;
-        try {
             lastLineReceived = "";
+            String response = null;
             writeString(commandAsString + "\r");
             long timeStamp = System.currentTimeMillis();
             while (lastLineReceived.equals("")) {
@@ -283,15 +297,125 @@ public class OceanicThingHandler extends SerialThingHandler {
                 }
                 if (System.currentTimeMillis() - timeStamp > REQUEST_TIMEOUT) {
                     logger.warn("A timeout occurred while requesting data from the water softener");
+                    readerThread.reset();
                     break;
                 }
             }
             response = lastLineReceived;
-        } finally {
+
             SerialPortThrottler.unlock(port);
+
+            return response;
+        }
+    }
+
+    public void writeString(String msg) {
+        String port = (String) this.getConfig().get(PORT);
+
+        try {
+            outputStream.write(msg.getBytes());
+            outputStream.flush();
+        } catch (IOException e) {
+            updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.COMMUNICATION_ERROR,
+                    "Error writing '" + msg + "' to serial port " + port + " : " + e.getMessage());
+        }
+    }
+
+    public class SerialPortReader extends Thread {
+
+        private boolean interrupted = false;
+        private InputStream inputStream;
+        int index = 0;
+
+        SimpleDateFormat sdf = new SimpleDateFormat("HH:mm:ss,SSS");
+
+        public SerialPortReader(InputStream in) {
+            this.inputStream = in;
+            this.setName("SerialPortReader-" + getThing().getUID());
         }
 
-        return response;
+        public void reset() {
+            index = 0;
+        }
 
+        @Override
+        public void interrupt() {
+            interrupted = true;
+            super.interrupt();
+            try {
+                inputStream.close();
+            } catch (IOException e) {
+                // quietly clos
+            }
+        }
+
+        @Override
+        public void run() {
+
+            byte[] dataBuffer = new byte[bufferSize];
+            byte[] tmpData = new byte[bufferSize];
+            boolean foundStart = false;
+            String line;
+
+            final byte lineFeed = (byte) '\n';
+            final byte carriageReturn = (byte) '\r';
+            final byte nullChar = (byte) '\0';
+
+            long sleep = 100;
+            int len = -1;
+
+            try {
+                while (interrupted != true) {
+                    if ((len = inputStream.read(tmpData)) > -1) {
+                        foundStart = false;
+                        for (int i = 0; i < len; i++) {
+                            if (i > 0) {
+                                if (tmpData[i] != lineFeed && tmpData[i] != carriageReturn
+                                        && tmpData[i] != nullChar) {
+                                    if (tmpData[i - 1] == lineFeed || tmpData[i - 1] == carriageReturn
+                                            || tmpData[i - 1] == nullChar) {
+                                        index = 0;
+                                        foundStart = true;
+                                    }
+                                }
+                            }
+
+                            if (tmpData[i] != lineFeed && tmpData[i] != carriageReturn && tmpData[i] != nullChar) {
+                                if (i == 0) {
+                                    foundStart = true;
+                                }
+                                dataBuffer[index++] = tmpData[i];
+                            }
+
+                            if (i > 0 && (tmpData[i] == lineFeed || tmpData[i] == carriageReturn
+                                    || tmpData[i] == nullChar)) {
+                                if (index > 0) {
+                                    if (foundStart) {
+                                        line = StringUtils.chomp(new String(Arrays.copyOf(dataBuffer, index)));
+                                        line = line.replace(",", ".");
+                                        line = line.trim();
+                                        lastLineReceived = line;
+                                        index = 0;
+                                        break;
+                                    }
+                                }
+                            }
+
+                            if (index == bufferSize) {
+                                index = 0;
+                            }
+                        }
+                    }
+
+                    try {
+                        Thread.sleep(sleep);
+                    } catch (InterruptedException e) {
+                        // Move on silently
+                    }
+                }
+            } catch (Exception e) {
+                logger.error("An exception occurred while reading serial port  : {}", e.getMessage(), e);
+            }
+        }
     }
 }
